[level 1] Title: 정수 내림차순으로 배치하기, Time: 0.97 ms, Memory: 84.5 MB -BaekjoonHub

+++ 프로그래머스/1/12933. 정수 내림차순으로 배치하기/정수 내림차순으로 배치하기.java
@@ -7,14 +7,14 @@ class Solution {
         List<Long> list = new ArrayList<>();
         
         while(n>0){
-            list.add(n % 10); 
+            list.add( n % 10); 
             n /= 10;
         }
         
         Collections.sort(list,Collections.reverseOrder());
         
         for(long a : list){
-            answer = answer*10 + a;
+            answer = (long)answer * 10 + a;
         }
